Refactored code of Map and UseMap
